🐛build fix: daemonize error when building multi platform distribution
